DS_SpA_W06_Dockerized_Data_Pipeline/02_19/murat_docker_comp/slackbot/requirements.txt
sqlalchemy
psycopg2-binary
slackclient
requests
pandas